rename state file
