+++ src/lib/components/initialize-game/utils/initializeBoard.ts
@@ -1,5 +1,5 @@
 import type { IBoard, IColumn } from "$lib/components/game/types";
-import { getGameState  } from "../../../state/state.svelte";
+import { getGameState  } from "../../../state/game.svelte";
 
 
 function makeSquareId (x: number, y: number, columns: number): number {
