+++ src/lib/state/game.svelte.ts
