+++ src/routes/+layout.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { setPlayerState } from '$lib/state/player.svelte';
-	import { setGameState } from '$lib/state/state.svelte';
+	import { setGameState } from '$lib/state/game.svelte';
 	import '../app.css';
 
 	let { children } = $props();
